build(Gradle): Move the Python package managers to their own plugin project
This continues the effort started in be4bb69 to move plugin
implementations, here for the Python package managers. See [1] for more
context.

[1]: https://github.com/oss-review-toolkit/ort/issues/6602
