+++ .gitmodules
@@ -1,9 +1,9 @@
 [submodule "example-python-flask"]
-	path = analyzer/src/funTest/assets/projects/external/example-python-flask
+	path = plugins/package-managers/python/src/funTest/assets/projects/external/example-python-flask
 	url = https://github.com/deis/example-python-flask.git
 	ignore = untracked
 [submodule "spdx-tools-python"]
-	path = analyzer/src/funTest/assets/projects/external/spdx-tools-python
+	path = plugins/package-managers/python/src/funTest/assets/projects/external/spdx-tools-python
 	url = https://github.com/spdx/tools-python.git
 	ignore = untracked
 [submodule "quickcheck-state-machine"]
