+++ analyzer/src/main/resources/META-INF/services/org.ossreviewtoolkit.analyzer.PackageManagerFactory
@@ -12,10 +12,7 @@ org.ossreviewtoolkit.analyzer.managers.Gradle$Factory
 org.ossreviewtoolkit.analyzer.managers.Maven$Factory
 org.ossreviewtoolkit.analyzer.managers.Npm$Factory
 org.ossreviewtoolkit.analyzer.managers.NuGet$Factory
-org.ossreviewtoolkit.analyzer.managers.Pip$Factory
-org.ossreviewtoolkit.analyzer.managers.Pipenv$Factory
 org.ossreviewtoolkit.analyzer.managers.Pnpm$Factory
-org.ossreviewtoolkit.analyzer.managers.Poetry$Factory
 org.ossreviewtoolkit.analyzer.managers.Pub$Factory
 org.ossreviewtoolkit.analyzer.managers.Sbt$Factory
 org.ossreviewtoolkit.analyzer.managers.SpdxDocumentFile$Factory
