+++ downloader/src/funTest/kotlin/vcs/GitWorkingTreeFunTest.kt
@@ -86,10 +86,10 @@ class GitWorkingTreeFunTest : StringSpec({
     "Git correctly lists submodules" {
         val expectedSubmodules = listOf(
             "analyzer/src/funTest/assets/projects/external/dart-http",
-            "analyzer/src/funTest/assets/projects/external/example-python-flask",
             "analyzer/src/funTest/assets/projects/external/quickcheck-state-machine",
             "analyzer/src/funTest/assets/projects/external/sbt-multi-project-example",
-            "analyzer/src/funTest/assets/projects/external/spdx-tools-python"
+            "plugins/package-managers/python/src/funTest/assets/projects/external/example-python-flask",
+            "plugins/package-managers/python/src/funTest/assets/projects/external/spdx-tools-python"
         ).associateWith { VersionControlSystem.getPathInfo(File("../$it")) }
 
         val workingTree = git.getWorkingTree(File(".."))
