+++ plugins/package-managers/python/src/funTest/assets/projects/external/example-python-flask-expected-output-windows.yml
