+++ plugins/package-managers/python/src/funTest/assets/projects/external/spdx-tools-python-expected-output.yml
