+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/pip-python3/django-project/settings.py
