+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/pip/setup.py
