+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/pipenv-python3-expected-output.yml
@@ -1,7 +1,7 @@
 ---
 project:
   id: "Pipenv::pipenv-python3-from-pipenv:<REPLACE_REVISION>"
-  definition_file_path: "analyzer/src/funTest/assets/projects/synthetic/pipenv-python3/requirements-from-pipenv.txt"
+  definition_file_path: "plugins/package-managers/python/src/funTest/assets/projects/synthetic/pipenv-python3/requirements-from-pipenv.txt"
   declared_licenses: []
   declared_licenses_processed: {}
   vcs:
