+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/pipenv-python3/django-project/wsgi.py
