+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/poetry/pyproject.toml
