+++ plugins/package-managers/python/src/funTest/assets/projects/synthetic/python-inspector/requirements.txt
