+++ plugins/package-managers/python/src/main/kotlin/utils/PythonInspector.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.analyzer.managers.utils
+package org.ossreviewtoolkit.plugins.packagemanagers.python.utils
 
 import java.io.File
 import java.util.SortedSet
